Moved view code out of VCs to separate view files
