+++ Catfish/Catfish/ViewControllers/CreateProfileViewController.swift
@@ -13,104 +13,33 @@ class CreateProfileViewController: FormViewController {
     // MARK: - View Lifecycle
     
     override func viewDidLoad() {
+        contentView = createProfileView
         super.viewDidLoad()
         configure()
     }
     
+    override func loadView() {
+        view = BGGradientView()
+    }
+    
     // MARK: - Private Properties
     
-    private var textViewPlaceholderManager: TextViewPlaceholderManager?
+    private var bioPlaceholderManager: TextViewPlaceholderManager?
     
-    // Views:
+    private let createProfileView = CreateProfileView()
     
-    private let createProfileLabel = UILabel(text: "Create your Catfish profile", font: Fonts.header, textAlignment: .center)
-    
-    private let nameLabel = UILabel(text: "Name:")
-    private let nameTextField = UITextField(placeholder: "Name")
-    private lazy var nameVStack: UIStackView = {
-        let nameVStack = UIStackView(arrangedSubviews: [nameLabel, nameTextField])
-        nameVStack.axis = .vertical
-        nameVStack.spacing = 10
-        return nameVStack
-    }()
-    
-    private let profilePicImageView: CircularImageView = {
-        let profilePic = CircularImageView(width: 150, image: UIImage(systemName: "person.circle.fill"))
-        profilePic.tintColor = Colors.purple
-        return profilePic
-    }()
-    private let uploadButton: UIButton = {
-        let uploadButton = UIButton(type: .system)
-        uploadButton.setTitle("Upload your Catfish profile picture", for: .normal)
-        uploadButton.setTitleColor(Colors.purple, for: .normal)
-        uploadButton.titleLabel?.font = .systemFont(ofSize: 16)
-        uploadButton.addTarget(self, action: #selector(uploadButtonTapped), for: .touchUpInside)
-        return uploadButton
-    }()
-    private lazy var profilePicVStack: UIStackView = {
-        let profilePicVStack = UIStackView(arrangedSubviews: [profilePicImageView, uploadButton])
-        profilePicVStack.axis = .vertical
-        profilePicVStack.alignment = .center
-        
-        return profilePicVStack
-    }()
-    
-    private let bioLabel = UILabel(text: "Bio:")
-    private let bioPlaceholder = "Say something about yourself..."
-    private lazy var bioTextView: UITextView = {
-        let bioTextView = UITextView()
-        bioTextView.translatesAutoresizingMaskIntoConstraints = false
-        bioTextView.heightAnchor.constraint(equalToConstant: 220).isActive = true
-        
-        bioTextView.layer.borderColor = UIColor.systemGray4.cgColor
-        bioTextView.layer.borderWidth = 1.0
-        bioTextView.layer.cornerRadius = 5.0
-        
-        textViewPlaceholderManager = TextViewPlaceholderManager(textView: bioTextView, placeholder: bioPlaceholder)
-        return bioTextView
-    }()
-    private lazy var bioVStack: UIStackView = {
-        let bioVStack = UIStackView(arrangedSubviews: [bioLabel, bioTextView])
-        bioVStack.axis = .vertical
-        bioVStack.spacing = 10
-        return bioVStack
-    }()
-    
-    private lazy var mainVStack: UIStackView = {
-        let mainVStack = UIStackView(arrangedSubviews: [
-            createProfileLabel,
-            profilePicVStack,
-            nameVStack,
-            bioVStack
-        ])
-        
-        mainVStack.axis = .vertical
-        mainVStack.spacing = 25
-        mainVStack.isLayoutMarginsRelativeArrangement = true
-        mainVStack.directionalLayoutMargins = .init(top: 20, leading: 20, bottom: 0, trailing: 20)
-        
-        return mainVStack
-    }()
-   
-    private let catfishButton = CFButton(backgroundColor: Colors.purple, title: "Catfish!", image: nil)
+    private lazy var profilePicImageView = createProfileView.profilePicImageView
+    private lazy var uploadButton = createProfileView.uploadButton
+    private lazy var nameTextField = createProfileView.nameTextField
+    private lazy var bioTextView = createProfileView.bioTextView
+    private lazy var catfishButton = createProfileView.catfishButton
+  
     
     // MARK: - Private Methods
     
     private func configure() {
-        view.backgroundColor = .white
-        
-        mainVStack.translatesAutoresizingMaskIntoConstraints = false
-        catfishButton.translatesAutoresizingMaskIntoConstraints = false
-        contentView.addSubviews(mainVStack, catfishButton)
-        
-        NSLayoutConstraint.activate([
-            mainVStack.topAnchor.constraint(equalTo: contentView.topAnchor),
-            mainVStack.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            mainVStack.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
-            
-            catfishButton.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
-            catfishButton.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -20)
-        ])
+        bioPlaceholderManager = TextViewPlaceholderManager(textView: bioTextView, placeholder: "Say something about yourself...")
+        uploadButton.addTarget(self, action: #selector(uploadButtonTapped), for: .touchUpInside)
     }
     
     // MARK: - Actions
